// backend/src/common/enums.ts
export enum Role {
  Admin = 'ADMIN',
  Participant = 'PARTICIPANT',
}

export enum EventStatus {
  Draft = 'DRAFT',
  Published = 'PUBLISHED',
  Canceled = 'CANCELED',
}

export enum ReservationStatus {
  Pending = 'PENDING',
  Confirmed = 'CONFIRMED',
  Refused = 'REFUSED',
  Canceled = 'CANCELED',
}